+++ tests/listing/test_basics.py
@@ -5,14 +5,14 @@ from tests.util import fixture_dir
 @pytest.mark.parametrize('outopt', [None, '-'])
 def test_output_no_file(capsys, outopt):
     search = fixture_dir('listing', 'valid-json')
-    npc.commands.listing(search, outfile=outopt)
+    npc.commands.listing.make_list(search, outfile=outopt)
     output, _ = capsys.readouterr()
     assert output
 
 def test_output_to_file(tmpdir):
     outfile = tmpdir.join("output.json")
     search = fixture_dir('listing', 'valid-json')
-    npc.commands.listing(search, outfile=str(outfile))
+    npc.commands.listing.make_list(search, outfile=str(outfile))
     assert outfile.read()
 
 def test_list_valid_json(list_json_output):